+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fix between placement (`SHIFT + X`) with control elements without `@startid`. 
 * Various GitHub fixes, incl. loading compressed MusicXML via Git
 * Check time stamps when inserting control elements, through warning for slur etc.
+* Unselect element(s) when `CTRL/CMD` clicking or drag-selecting already selected element(s)
 
 ### 0.8.5 PDF export
 * Preview PDF mode with page-view of notation, triggered via menu item or `CMD/CTRL + P`