+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # mei-friend-online CHANGELOG.md
+### 0.8.7 transposition
+* Add transposition settings panel to the mei-friend settings to specify transposition by key, by interval, and pitch direction.
+  
 ### 0.8.6 patch
 * Insert control elements with time stamps (`@tstamp`/`tstamp2`) instead of ids (`startid`/`endid`) by using the  `ALT` (Mac: `CTRL`) modifier key with keyboard shortcut
 * Check time stamps when inserting control elements, handle selected elements on the same time stamp correctly (by omitting `@tstamp2`) or throw warning, if no fix is possible